+++ models/psat_pl_perubahan.js
@@ -20,7 +20,7 @@ class PsatPlPerubahanModel {
         let file_permohonan, perubahan_data;
         try {
             let response = {};
-
+            await check_query.check_data(data)
             //Create new file pemohonan
             let data_file_permohonan = [
                 data.id_pengguna, data.surat_permohonan_izin_edar, data.sertifikat_izin_edar_sebelumnya,
@@ -47,6 +47,9 @@ class PsatPlPerubahanModel {
             // debug('get %o', response);
             return { status: '200', permohohan: "Perubahan Data Izin Edar PSAT PL", notifikasi: notif, data: response };
         } catch (ex) {
+            if(ex.code == '401'){
+                return { status: '400', Error: ex.pesan };
+            }
             let delete_pengajuan = await pool.query('DELETE FROM ' + db_pengajuan + ' WHERE id = $1 RETURNING *', [perubahan_data.rows[0].id]);
             await pool.query('DELETE FROM ' + db_file_permohonan + ' WHERE id = $1 RETURNING *', [delete_pengajuan.rows[0].file_permohonan]);
             console.log('Enek seng salah iki ' + ex);
@@ -79,6 +82,7 @@ class PsatPlPerubahanModel {
 
     async add_perubahan_info_produk(data) {
         try {
+            await check_query.check_data(data)
             let data_info_produk = [
                 data.id_pengguna, data.jenis_psat, data.nama_latin, data.nama_dagang, data.nomor_sertifikat,
                 data.expire_sertifikat, data.file_sertifikat, data.data_awal, data.data_baru,
@@ -92,6 +96,9 @@ class PsatPlPerubahanModel {
             // debug('get %o', res);
             return { status: '200', permohohan: "Add Perubahan Info Produk", data: info_produk.rows[0] };
         } catch (ex) {
+            if(ex.code == '401'){
+                return { status: '400', Error: ex.pesan };
+            }
             console.log('Enek seng salah iki ' + ex);
             return { status: '400', Error: "" + ex };
         };
@@ -101,7 +108,7 @@ class PsatPlPerubahanModel {
         try {
             let response = {};
             let file_permohonan, perubahan_data;
-
+            await check_query.check_data(data)
             //Update pengajuan
             let data_perubahan_data = [true, data.status_pengajuan, 10, date]
             perubahan_data = await pool.query(
@@ -128,6 +135,9 @@ class PsatPlPerubahanModel {
             // debug('get %o', response);
             return { status: '200', permohohan: "Update Perubahan Data Izin Edar PSAT PL", data: response };
         } catch (ex) {
+            if(ex.code == '401'){
+                return { status: '400', Error: ex.pesan };
+            }
             console.log('Enek seng salah iki ' + ex);
             return { status: '400', Error: "" + ex };
         };
@@ -158,6 +168,7 @@ class PsatPlPerubahanModel {
 
     async update_perubahan_info_produk(data) {
         try {
+            await check_query.check_data(data)
             let data_info_produk = [
                 data.jenis_psat, data.nama_latin, data.nama_dagang, data.nomor_sertifikat,
                 data.expire_sertifikat, data.file_sertifikat, data.data_awal, data.data_baru,
@@ -172,6 +183,9 @@ class PsatPlPerubahanModel {
             // debug('get %o', res);
             return { status: '200', permohohan: "Update Perubahan Info Produk", data: info_produk.rows[0] };
         } catch (ex) {
+            if(ex.code == '401'){
+                return { status: '400', Error: ex.pesan };
+            }s
             console.log('Enek seng salah iki ' + ex);
             return { status: '400', Error: "" + ex };
         };
